requirements.txt
Brotli==1.0.9
certifi==2025.4.26
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
gmpy2==2.2.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.3
onnx==1.17.0
onnxruntime-gpu==1.22.0
onnxslim==0.1.52
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.1.0
pip==25.1
protobuf==6.30.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
sympy==1.13.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.119
ultralytics-thop==2.0.14
urllib3==2.3.0
wheel==0.45.1
win-inet-pton==1.1.0
